refactor(test): reorganize tests into unit and integration classes
Establish consistent test naming convention:
- *ServiceTest.java = unit tests (mocked dependencies, fast)
- *ServiceIntegrationTest.java = integration tests (Testcontainers, real Solr)

Search tests:
- SearchServiceTest: unit tests with mocked SolrClient (replaces SearchServiceDirectTest)
- SearchServiceIntegrationTest: real Solr via Testcontainers (new)

Indexing tests:
- IndexingServiceTest: unit tests with mocked SolrClient (replaces nested UnitTests + DirectTest)
- IndexingServiceIntegrationTest: real Solr via Testcontainers (new)

Also:
- Delete redundant *DirectTest.java files
- Fix TestcontainersConfiguration default Solr image to "solr:9.9-slim"
